requirements.txt
Flask==1.0.2
Pillow==5.3.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyTelegramBotAPI==3.6.6
gunicorn==19.9.0
